Cleanup

+++ DgTemp-Firmware/DgTemp-Firmware.ino
@@ -4,7 +4,6 @@
 #include "src/TeensyDAC/TeensyDAC.h"
 #include <DMAChannel.h>
 
-
 #define CS_DAC          8
 AD5760 dac(CS_DAC);
 TeensyDAC DCDC;
@@ -16,10 +15,7 @@ TeensyDAC DCDC;
  * 10: Inverted
  */
 
-
 const float ADC_V_REF = 4.096;
-  
-
 
 void setup(){
   clockInit();
@@ -36,29 +32,20 @@ void setup(){
   delay(500);                               // Deleay needed between setting DAC_CS-Pin and the first SPI-Transaction (TODO: test for minimum duration)
   dac.reset();                              // executes enableOutput
   dac.setValue(39999);                      // DAC Setpoint is set
-
   
   spiInit();
   timerInit();
 
   Serial.begin(115200);
   Serial.println("Setup Worked");
-
-
-
-  
-  
 }
-
-
 
 void loop(){  
   // Flush serial buffer to submit data to Linux systems
-
- 
   while (Serial.available()) {
     Serial.read();
   }
+
   if(newSample) {
     Serial.print((int32_t)code);
     //Serial.print(codeToVoltage(code,ADC_V_REF),10);
@@ -72,23 +59,8 @@ void loop(){
     
     newSample = false;
   }
-
-  
 }
 
-
-
-
-
-
-float codeToVoltage(const int32_t code, const float vref) {
-  return ((float)code / 2147483647) * vref;
+double codeToVoltage(const int32_t code, const float vref) {
+  return ((double)code / 2147483647) * vref;
 }
-
-
-
-  
-
-
-
-
